+++ src/main/java/nl/tudelft/skills/dto/view/edition/EditionLevelEditionViewDTO.java
@@ -55,12 +55,15 @@ public class EditionLevelEditionViewDTO extends View<SCEdition> implements Circu
 	@NotNull
 	@PostApply
 	private List<CheckpointViewDTO> checkpointsInEdition;
+	@NotNull
+	private Set<PathViewDTO> paths;
 
 	@Override
 	public List<? extends GroupView> getGroups() {
 		return modules;
 	}
 
+	@Override
 	public void postApply() {
 		super.postApply();
 		this.checkpointsInEdition = View.convert(